backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Redis
redis==5.0.1
hiredis==2.3.2

# Google Cloud
google-cloud-storage==2.14.0
google-cloud-kms==2.20.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.116.0

# Document Processing
PyPDF2==3.0.1
pdfplumber==0.10.3  # Best PDF text extraction - WORKS GREAT
pypdf==3.17.4  # Alternative PDF library
python-docx==1.1.0
openpyxl==3.1.2
pandas==2.2.0
pymupdf==1.23.8  # PyMuPDF - excellent for PDF extraction, faster than PyPDF2
# tabula-py==2.9.0  # Table extraction - requires Java, skip for now
# camelot-py[cv]==0.11.0  # Table extraction - installation issues, skip for now

# AI/ML and NLP
langchain==0.1.4
langchain-community==0.0.16
langchain-openai==0.0.5
openai==1.10.0
anthropic==0.8.1
tiktoken==0.5.2
sentence-transformers==2.3.1
transformers==4.37.2
torch==2.2.0

# Vector Database
pinecone-client==3.0.2
chromadb==0.4.22

# Security
cryptography==42.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
pyotp==2.9.0

# Utilities
python-multipart==0.0.6
python-dateutil==2.8.2
pytz==2024.1
requests==2.31.0
httpx==0.26.0
aiofiles==23.2.1
celery==5.3.6
python-dotenv==1.0.1

# Monitoring and Logging
sentry-sdk==1.40.0
structlog==24.1.0
python-json-logger==2.0.7

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.26.0
faker==22.5.1

# Development
black==24.1.1
flake8==7.0.0
mypy==1.8.0
isort==5.13.2
pre-commit==3.6.0

# Email
fastapi-mail==1.4.1

# Rate Limiting
slowapi==0.1.9

# Documentation
mkdocs==1.5.3
mkdocs-material==9.5.6
